+++ datanode/gateway/graphql/generated.go
@@ -1118,10 +1118,11 @@ type ComplexityRoot struct {
 	}
 
 	ProposalVoteSide struct {
-		TotalNumber func(childComplexity int) int
-		TotalTokens func(childComplexity int) int
-		TotalWeight func(childComplexity int) int
-		Votes       func(childComplexity int) int
+		TotalEquityLikeShareWeight func(childComplexity int) int
+		TotalNumber                func(childComplexity int) int
+		TotalTokens                func(childComplexity int) int
+		TotalWeight                func(childComplexity int) int
+		Votes                      func(childComplexity int) int
 	}
 
 	ProposalVotes struct {
@@ -1518,6 +1519,7 @@ type ComplexityRoot struct {
 
 	Vote struct {
 		Datetime               func(childComplexity int) int
+		EquityLikeShareWeight  func(childComplexity int) int
 		GovernanceTokenBalance func(childComplexity int) int
 		GovernanceTokenWeight  func(childComplexity int) int
 		Party                  func(childComplexity int) int
@@ -2065,6 +2067,7 @@ type VoteResolver interface {
 
 	GovernanceTokenBalance(ctx context.Context, obj *vega.Vote) (string, error)
 	GovernanceTokenWeight(ctx context.Context, obj *vega.Vote) (string, error)
+	EquityLikeShareWeight(ctx context.Context, obj *vega.Vote) (string, error)
 }
 type WithdrawalResolver interface {
 	Party(ctx context.Context, obj *vega.Withdrawal) (*vega.Party, error)
@@ -6419,6 +6422,13 @@ func (e *executableSchema) Complexity(typeName, field string, childComplexity in
 
 		return e.complexity.ProposalVoteEdge.Node(childComplexity), true
 
+	case "ProposalVoteSide.totalEquityLikeShareWeight":
+		if e.complexity.ProposalVoteSide.TotalEquityLikeShareWeight == nil {
+			break
+		}
+
+		return e.complexity.ProposalVoteSide.TotalEquityLikeShareWeight(childComplexity), true
+
 	case "ProposalVoteSide.totalNumber":
 		if e.complexity.ProposalVoteSide.TotalNumber == nil {
 			break
@@ -8518,6 +8528,13 @@ func (e *executableSchema) Complexity(typeName, field string, childComplexity in
 		}
 
 		return e.complexity.Vote.Datetime(childComplexity), true
+
+	case "Vote.equityLikeShareWeight":
+		if e.complexity.Vote.EquityLikeShareWeight == nil {
+			break
+		}
+
+		return e.complexity.Vote.EquityLikeShareWeight(childComplexity), true
 
 	case "Vote.governanceTokenBalance":
 		if e.complexity.Vote.GovernanceTokenBalance == nil {
@@ -11989,6 +12006,8 @@ type ProposalVoteSide {
   totalWeight: String!
   "Total number of governance tokens from the votes cast for this side"
   totalTokens: String!
+  "Total equity like share weight for this side (only for UpdateMarket Proposals)"
+  totalEquityLikeShareWeight: String!
 }
 
 enum VoteValue {
@@ -12016,6 +12035,9 @@ type Vote {
 
   "The weight of this vote based on the total of governance token"
   governanceTokenWeight: String!
+
+  "The weight of this vote based on the total equity like share"
+  equityLikeShareWeight: String!
 }
 
 type ProposalVote {
@@ -36072,6 +36094,41 @@ func (ec *executionContext) _ProposalVoteSide_totalTokens(ctx context.Context, f
 	return ec.marshalNString2string(ctx, field.Selections, res)
 }
 
+func (ec *executionContext) _ProposalVoteSide_totalEquityLikeShareWeight(ctx context.Context, field graphql.CollectedField, obj *ProposalVoteSide) (ret graphql.Marshaler) {
+	defer func() {
+		if r := recover(); r != nil {
+			ec.Error(ctx, ec.Recover(ctx, r))
+			ret = graphql.Null
+		}
+	}()
+	fc := &graphql.FieldContext{
+		Object:     "ProposalVoteSide",
+		Field:      field,
+		Args:       nil,
+		IsMethod:   false,
+		IsResolver: false,
+	}
+
+	ctx = graphql.WithFieldContext(ctx, fc)
+	resTmp, err := ec.ResolverMiddleware(ctx, func(rctx context.Context) (interface{}, error) {
+		ctx = rctx // use context from middleware stack in children
+		return obj.TotalEquityLikeShareWeight, nil
+	})
+	if err != nil {
+		ec.Error(ctx, err)
+		return graphql.Null
+	}
+	if resTmp == nil {
+		if !graphql.HasFieldError(ctx, fc) {
+			ec.Errorf(ctx, "must not be null")
+		}
+		return graphql.Null
+	}
+	res := resTmp.(string)
+	fc.Result = res
+	return ec.marshalNString2string(ctx, field.Selections, res)
+}
+
 func (ec *executionContext) _ProposalVotes_yes(ctx context.Context, field graphql.CollectedField, obj *ProposalVotes) (ret graphql.Marshaler) {
 	defer func() {
 		if r := recover(); r != nil {
@@ -45398,6 +45455,41 @@ func (ec *executionContext) _Vote_governanceTokenWeight(ctx context.Context, fie
 	resTmp, err := ec.ResolverMiddleware(ctx, func(rctx context.Context) (interface{}, error) {
 		ctx = rctx // use context from middleware stack in children
 		return ec.resolvers.Vote().GovernanceTokenWeight(rctx, obj)
+	})
+	if err != nil {
+		ec.Error(ctx, err)
+		return graphql.Null
+	}
+	if resTmp == nil {
+		if !graphql.HasFieldError(ctx, fc) {
+			ec.Errorf(ctx, "must not be null")
+		}
+		return graphql.Null
+	}
+	res := resTmp.(string)
+	fc.Result = res
+	return ec.marshalNString2string(ctx, field.Selections, res)
+}
+
+func (ec *executionContext) _Vote_equityLikeShareWeight(ctx context.Context, field graphql.CollectedField, obj *vega.Vote) (ret graphql.Marshaler) {
+	defer func() {
+		if r := recover(); r != nil {
+			ec.Error(ctx, ec.Recover(ctx, r))
+			ret = graphql.Null
+		}
+	}()
+	fc := &graphql.FieldContext{
+		Object:     "Vote",
+		Field:      field,
+		Args:       nil,
+		IsMethod:   true,
+		IsResolver: true,
+	}
+
+	ctx = graphql.WithFieldContext(ctx, fc)
+	resTmp, err := ec.ResolverMiddleware(ctx, func(rctx context.Context) (interface{}, error) {
+		ctx = rctx // use context from middleware stack in children
+		return ec.resolvers.Vote().EquityLikeShareWeight(rctx, obj)
 	})
 	if err != nil {
 		ec.Error(ctx, err)
@@ -58083,6 +58175,16 @@ func (ec *executionContext) _ProposalVoteSide(ctx context.Context, sel ast.Selec
 			if out.Values[i] == graphql.Null {
 				invalids++
 			}
+		case "totalEquityLikeShareWeight":
+			innerFunc := func(ctx context.Context) (res graphql.Marshaler) {
+				return ec._ProposalVoteSide_totalEquityLikeShareWeight(ctx, field, obj)
+			}
+
+			out.Values[i] = innerFunc(ctx)
+
+			if out.Values[i] == graphql.Null {
+				invalids++
+			}
 		default:
 			panic("unknown field " + strconv.Quote(field.Name))
 		}
@@ -62683,6 +62785,26 @@ func (ec *executionContext) _Vote(ctx context.Context, sel ast.SelectionSet, obj
 					}
 				}()
 				res = ec._Vote_governanceTokenWeight(ctx, field, obj)
+				if res == graphql.Null {
+					atomic.AddUint32(&invalids, 1)
+				}
+				return res
+			}
+
+			out.Concurrently(i, func() graphql.Marshaler {
+				return innerFunc(ctx)
+
+			})
+		case "equityLikeShareWeight":
+			field := field
+
+			innerFunc := func(ctx context.Context) (res graphql.Marshaler) {
+				defer func() {
+					if r := recover(); r != nil {
+						ec.Error(ctx, ec.Recover(ctx, r))
+					}
+				}()
+				res = ec._Vote_equityLikeShareWeight(ctx, field, obj)
 				if res == graphql.Null {
 					atomic.AddUint32(&invalids, 1)
 				}